Some changes..
There was not much I was able to do for home. I still had a lot of trouble with getting the segments to get on (and with dissapointment I can conclude I still do not have the segments) and without the segments working I am unable to finish the food function for adding segments, the score for the segments and food and the game over sequence for when the head hits a wall, a lot of my code is messy so if I do have time in class I should probably start with cleaning that up first.