adding another segment.
finally managed to add another segment to the snake however I still find myself unable to add it each time the food is eaten. the array is filling up and I do now know how to call it normally but I do not understand where exactly to put the code so that the segments will be drawn each time and there is a small (but not code breaking) problem with the for function that travels through this array. hopefully this will be solved at home.