Additions to the game.
I, FINALLY managed to get the snake to increase in segments every time it ate food and added a small scoreboard. As much as I'd like to stay up past 12:22 I'm instead choosing to sleep. This means I have not finished the restarting game when the snake intersects itself or hits a wall. Besides this a lot has been added.